Rewrite ckeditor5-paste-from-office to TypeScript.
