+++ .gitignore
@@ -32,6 +32,7 @@ packages/ckeditor5-core/src/**/*.js
 packages/ckeditor5-engine/src/**/*.js
 packages/ckeditor5-enter/src/**/*.js
 packages/ckeditor5-paragraph/src/**/*.js
+packages/ckeditor5-paste-from-office/src/**/*.js
 packages/ckeditor5-select-all/src/**/*.js
 packages/ckeditor5-typing/src/**/*.js
 packages/ckeditor5-ui/src/**/*.js
