+++ packages/ckeditor5-clipboard/src/index.ts
@@ -8,6 +8,10 @@
  */
 
 export { default as Clipboard } from './clipboard';
-export { default as ClipboardPipeline } from './clipboardpipeline';
+export {
+	default as ClipboardPipeline,
+	type ClipboardInputTransformationEvent,
+	type ClipboardInputTransformationData
+} from './clipboardpipeline';
 export { default as DragDrop } from './dragdrop';
 export { default as PastePlainText } from './pasteplaintext';
