+++ packages/ckeditor5-engine/src/view/element.ts
@@ -677,17 +677,17 @@ export default class Element extends Node {
 	 * @param {String} value Attribute value.
 	 * @fires module:engine/view/node~Node#change
 	 */
-	public _setAttribute( key: string, value: string ): void {
-		value = String( value );
+	public _setAttribute( key: string, value: unknown ): void {
+		const stringValue = String( value );
 
 		this._fireChange( 'attributes', this );
 
 		if ( key == 'class' ) {
-			parseClasses( this._classes, value );
+			parseClasses( this._classes, stringValue );
 		} else if ( key == 'style' ) {
-			this._styles.setTo( value );
+			this._styles.setTo( stringValue );
 		} else {
-			this._attrs.set( key, value );
+			this._attrs.set( key, stringValue );
 		}
 	}
 
@@ -913,7 +913,7 @@ Element.prototype.is = function( type: string, name?: string ): boolean {
 	}
 };
 
-export type ElementAttributes = Record<string, string> | Iterable<[ string, string ]> | null;
+export type ElementAttributes = Record<string, unknown> | Iterable<[ string, unknown ]> | null;
 
 // Parses attributes provided to the element constructor before they are applied to an element. If attributes are passed
 // as an object (instead of `Iterable`), the object is transformed to the map. Attributes with `null` value are removed.
@@ -932,7 +932,7 @@ function parseAttributes( attrs?: ElementAttributes ) {
 		}
 	}
 
-	return attrsMap;
+	return attrsMap as Map<string, string>;
 }
 
 // Parses class attribute and puts all classes into classes set.
