+++ packages/ckeditor5-paste-from-office/src/filters/removeboldwrapper.ts
@@ -7,13 +7,14 @@
  * @module paste-from-office/filters/removeboldwrapper
  */
 
+import type { UpcastWriter, ViewDocumentFragment } from 'ckeditor5/src/engine';
+
 /**
  * Removes `<b>` tag wrapper added by Google Docs to a copied content.
  *
- * @param {module:engine/view/documentfragment~DocumentFragment} documentFragment element `data.content` obtained from clipboard
- * @param {module:engine/view/upcastwriter~UpcastWriter} writer
+ * @param documentFragment element `data.content` obtained from clipboard
  */
-export default function removeBoldWrapper( documentFragment, writer ) {
+export default function removeBoldWrapper( documentFragment: ViewDocumentFragment, writer: UpcastWriter ): void {
 	for ( const child of documentFragment.getChildren() ) {
 		if ( child.is( 'element', 'b' ) && child.getStyle( 'font-weight' ) === 'normal' ) {
 			const childIndex = documentFragment.getChildIndex( child );
