+++ packages/ckeditor5-paste-from-office/src/filters/space.ts
@@ -13,10 +13,10 @@
  * This method also takes into account Word specific `<o:p></o:p>` empty tags.
  * Additionally multiline sequences of spaces and new lines between tags are removed (see #39 and #40).
  *
- * @param {String} htmlString HTML string in which spacing should be normalized.
- * @returns {String} Input HTML with spaces normalized.
+ * @param htmlString HTML string in which spacing should be normalized.
+ * @returns Input HTML with spaces normalized.
  */
-export function normalizeSpacing( htmlString ) {
+export function normalizeSpacing( htmlString: string ): string {
 	// Run normalizeSafariSpaceSpans() two times to cover nested spans.
 	return normalizeSafariSpaceSpans( normalizeSafariSpaceSpans( htmlString ) )
 		// Remove all \r\n from "spacerun spans" so the last replace line doesn't strip all whitespaces.
@@ -35,26 +35,29 @@ export function normalizeSpacing( htmlString ) {
  * all spaces with `&nbsp; ` pairs. This prevents spaces from being removed during further DOM/View processing
  * (see especially {@link module:engine/view/domconverter~DomConverter#_processDataFromDomText}).
  *
- * @param {Document} htmlDocument Native `Document` object in which spacing should be normalized.
+ * @param htmlDocument Native `Document` object in which spacing should be normalized.
  */
-export function normalizeSpacerunSpans( htmlDocument ) {
+export function normalizeSpacerunSpans( htmlDocument: Document ): void {
 	htmlDocument.querySelectorAll( 'span[style*=spacerun]' ).forEach( el => {
-		const innerTextLength = el.innerText.length || 0;
+		const htmlElement = el as HTMLElement;
+		const innerTextLength = htmlElement.innerText.length || 0;
 
-		el.innerText = Array( innerTextLength + 1 ).join( '\u00A0 ' ).substr( 0, innerTextLength );
+		htmlElement.innerText = Array( innerTextLength + 1 ).join( '\u00A0 ' ).substr( 0, innerTextLength );
 	} );
 }
 
-// Normalizes specific spacing generated by Safari when content pasted from Word (`<span class="Apple-converted-space"> </span>`)
-// by replacing all spaces sequences longer than 1 space with `&nbsp; ` pairs. This prevents spaces from being removed during
-// further DOM/View processing (see especially {@link module:engine/view/domconverter~DomConverter#_processDataFromDomText}).
-//
-// This function is similar to {@link module:clipboard/utils/normalizeclipboarddata normalizeClipboardData util} but uses
-// regular spaces / &nbsp; sequence for replacement.
-//
-// @param {String} htmlString HTML string in which spacing should be normalized
-// @returns {String} Input HTML with spaces normalized.
-function normalizeSafariSpaceSpans( htmlString ) {
+/**
+ * Normalizes specific spacing generated by Safari when content pasted from Word (`<span class="Apple-converted-space"> </span>`)
+ * by replacing all spaces sequences longer than 1 space with `&nbsp; ` pairs. This prevents spaces from being removed during
+ * further DOM/View processing (see especially {@link module:engine/view/domconverter~DomConverter#_processDataFromDomText}).
+ *
+ * This function is similar to {@link module:clipboard/utils/normalizeclipboarddata normalizeClipboardData util} but uses
+ * regular spaces / &nbsp; sequence for replacement.
+ *
+ * @param htmlString HTML string in which spacing should be normalized
+ * @returns Input HTML with spaces normalized.
+ */
+function normalizeSafariSpaceSpans( htmlString: string ) {
 	return htmlString.replace( /<span(?: class="Apple-converted-space"|)>(\s+)<\/span>/g, ( fullMatch, spaces ) => {
 		return spaces.length === 1 ? ' ' : Array( spaces.length + 1 ).join( '\u00A0 ' ).substr( 0, spaces.length );
 	} );
