+++ packages/ckeditor5-paste-from-office/src/index.ts
