+++ packages/ckeditor5-paste-from-office/src/normalizers/googledocsnormalizer.ts
@@ -7,44 +7,41 @@
  * @module paste-from-office/normalizers/googledocsnormalizer
  */
 
-import { UpcastWriter } from 'ckeditor5/src/engine';
+import { UpcastWriter, type ViewDocument } from 'ckeditor5/src/engine';
 
 import removeBoldWrapper from '../filters/removeboldwrapper';
 import transformBlockBrsToParagraphs from '../filters/br';
 import { unwrapParagraphInListItem } from '../filters/list';
+import type { Normalizer, NormalizerData } from '../normalizer';
 
 const googleDocsMatch = /id=("|')docs-internal-guid-[-0-9a-f]+("|')/i;
 
 /**
  * Normalizer for the content pasted from Google Docs.
- *
- * @implements module:paste-from-office/normalizer~Normalizer
  */
-export default class GoogleDocsNormalizer {
+export default class GoogleDocsNormalizer implements Normalizer {
+	public readonly document: ViewDocument;
+
 	/**
 	 * Creates a new `GoogleDocsNormalizer` instance.
 	 *
-	 * @param {module:engine/view/document~Document} document View document.
+	 * @param document View document.
 	 */
-	constructor( document ) {
-		/**
-		 * @readonly
-		 * @type {module:engine/view/document~Document}
-		 */
+	constructor( document: ViewDocument ) {
 		this.document = document;
 	}
 
 	/**
 	 * @inheritDoc
 	 */
-	isActive( htmlString ) {
+	public isActive( htmlString: string ): boolean {
 		return googleDocsMatch.test( htmlString );
 	}
 
 	/**
 	 * @inheritDoc
 	 */
-	execute( data ) {
+	public execute( data: NormalizerData ): void {
 		const writer = new UpcastWriter( this.document );
 		const { body: documentFragment } = data._parsedData;
