+++ packages/ckeditor5-paste-from-office/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "ckeditor5/tsconfig.json",
+  "include": [
+    "src",
+    "../../typings"
+  ]
+}
